refactor: context() 메서드 JdbcTemplate 으로 이동
