+++ app/src/main/java/com/techcourse/dao/Strategy/FindByAccountStrategy.java
@@ -1,5 +1,7 @@
 package com.techcourse.dao.Strategy;
 
+import org.springframework.jdbc.core.PreparedStrategy;
+
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
