--- app/src/main/java/com/techcourse/dao/Strategy/UpdateStrategy.java
@@ -1,13 +1,0 @@
-package com.techcourse.dao.Strategy;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-public class UpdateStrategy implements PreparedStrategy {
-
-    @Override
-    public PreparedStatement createStatement(Connection connection) throws SQLException {
-        return connection.prepareStatement("update users set account = ?, password = ?, email = ? where id = ?");
-    }
-}
