+++ jdbc/src/main/java/org/springframework/jdbc/core/PreparedStrategy.java
@@ -1,4 +1,4 @@
-package com.techcourse.dao.Strategy;
+package org.springframework.jdbc.core;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
